change float into double precision.  the linear interpolation in resample function produce errors witch sounding like distortion. now no more distortion will produced.
git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@1253 3b3fb362-3133-0410-aa15-cf69e0a59cb7

+++ libs/hydrogen/src/adsr.cpp
@@ -27,7 +27,7 @@
 namespace H2Core
 {
 
-inline static float linear_interpolation( float fVal_A, float fVal_B, float fVal )
+inline static float linear_interpolation( float fVal_A, float fVal_B, double fVal )
 {
 	return fVal_A * ( 1 - fVal ) + fVal_B * fVal;
 //	return fVal_A + fVal * ( fVal_B - fVal_A );
